1. 배터리 정보 주는거 임시로 추가 2. rescan 시 죽는 현상 수정

+++ gpr_socket/gpr_socket_acq.c
@@ -112,9 +112,12 @@ void endFileWrite()
 
 void deleteAcqFile()
 {
-    char *path = getFullPath(acqCon.savePath, acqCon.fileName);
-    deleteFile(path);
-    free(path);
+    if (acqCon.savePath != NULL && acqCon.fileName != NULL)
+    {
+        char *path = getFullPath(acqCon.savePath, acqCon.fileName);
+        deleteFile(path);
+        free(path);
+    }
 }
 
 void deleteAcq3DFolder()
